# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hermes Studio API",
    "description": "Hermes Studio API — chat sessions, scheduled jobs, platform channels, model management, skills, memory, logs, file browser, group chat, and terminal.",
    "version": "0.6.15"
  },
  "servers": [
    {
      "url": "http://localhost:8648",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/devices/peer-connections/{connectionId}/download": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Create download",
        "description": "POST /api/devices/peer-connections/:connectionId/download",
        "operationId": "downloadPeerFile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "parameters": [
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/hermes/download": {
      "get": {
        "tags": [
          "Download"
        ],
        "summary": "Get download",
        "description": "GET /api/hermes/download",
        "operationId": "getDownload",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/hermes/performance/runtime": {
      "get": {
        "tags": [
          "Performance"
        ],
        "summary": "Get runtime",
        "description": "GET /api/hermes/performance/runtime",
        "operationId": "runtime",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/hermes/runtime-versions/runtime/download": {
      "post": {
        "tags": [
          "Runtime Versions"
        ],
        "summary": "Create download",
        "description": "POST /api/hermes/runtime-versions/runtime/download",
        "operationId": "downloadRuntime",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "version": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hermes/runtime-versions/webui/download": {
      "post": {
        "tags": [
          "Runtime Versions"
        ],
        "summary": "Create download",
        "description": "POST /api/hermes/runtime-versions/webui/download",
        "operationId": "downloadWebUi",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "version": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hermes/sessions/hermes": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get hermes",
        "description": "GET /api/hermes/sessions/hermes",
        "operationId": "listHermesSessions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/hermes/sessions/hermes/{id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get :id",
        "description": "GET /api/hermes/sessions/hermes/:id",
        "operationId": "getHermesSession",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request - Invalid parameters",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request"
                }
              }
            }
          }
        }
      },
...
    }
  }
}